backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "earthengine-api>=1.7.20",
    "fastapi>=0.135.3",
    "geopandas>=1.1.3",
    "numpy>=2.4.4",
    "pyproj>=3.7.2",
    "python-dotenv>=1.2.2",
    "rasterio>=1.5.0",
    "requests>=2.33.1",
    "scikit-image>=0.26.0",
    "shapely>=2.1.2",
    "uvicorn[standard]>=0.42.0",
]
